frameout bug fixed
"frameout" prints pdb files with the keyword MODEL followed by its
corresponding trajectory frame. The file is compatible with pymol and
now with nglview. The function in charge of printing ATOMS is now
embedded in the function WriteTimeFrame.

NOTE: "filter" program was changed to follow a similar behavior as
frameout and filter depend on the same class OutPdb_i